featI(one-bash-run-to-get-all-done): Add scripts to install SonarQube scanner
This commit adds a set of scripts to automate the installation of the SonarQube scanner on different platforms:

- index.sh - Main script that lets user select OS and executes the corresponding install script
- install_sonarqube-scanner_centos.sh - Installs scanner on CentOS
- install_sonarqube-scanner_ubuntu.sh - Installs scanner on Ubuntu
- install-sonar-scanner-macosx.sh - Installs scanner on Mac OSX
- install-sonar-scanner-windows-bash.sh - Installs on Windows (Git Bash)
- install-sonar-scanner-windows-powershell.ps1 - Installs on Windows (PowerShell)

The scripts handle downloading the correct scanner version, unzipping, moving to opt directory, and creating symlinks. This enables easy, automated installation of the scanner across environments.

Testing and verification has been done on all platforms.

This allows new developers to quickly get the scanner installed and makes it easy to add scanning to CI pipelines.